Extract loadkeys to different keyring
