+++ src/lib/keyring.c
@@ -1,0 +1,106 @@
+/*
+ * Copyright (c) 2017, [Ribose Inc](https://www.ribose.com).
+ * All rights reserved.
+ *
+ * This code is originally derived from software contributed to
+ * The NetBSD Foundation by Alistair Crooks (agc@netbsd.org), and
+ * carried further by Ribose Inc (https://www.ribose.com).
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE NETBSD FOUNDATION, INC. AND CONTRIBUTORS
+ * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE FOUNDATION OR CONTRIBUTORS
+ * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "config.h"
+
+#include "rnp.h"
+#include "keyring_pgp.h"
+#include "keyring_ssh.h"
+
+#include <stdio.h>
+#include <string.h>
+
+/* Format a PGP key to a readable hexadecimal string in a user supplied
+ * buffer.
+ *
+ * buffer: the buffer to write into
+ * sigid:  the PGP key ID to format
+ * len:    the length of buffer, including the null terminator
+ *
+ * TODO: There is no error checking here.
+ * TODO: Make this function more general or use an existing one.
+ */
+
+void
+keyring_format_key(char *buffer, uint8_t *sigid, int len)
+{
+    unsigned int i;
+    unsigned int n;
+
+    /* Chunks of two bytes are processed at a time because we can
+     * always be reasonably sure that PGP_KEY_ID_SIZE will be
+     * divisible by two. However, if the RFCs specify a fixed
+     * fixed size for PGP key IDs it might be more constructive
+     * to format this in one call and do a compile-time size
+     * check of the constant. If somebody wanted to do
+     * something exotic they can easily re-implement
+     * this function.
+     */
+    for (i = 0, n = 0; i < PGP_KEY_ID_SIZE; i += 2) {
+        n += snprintf(&buffer[n], len - n, "%02x%02x",
+                      sigid[i], sigid[i + 1]);
+    }
+    buffer[n] = 0x0;
+}
+
+/* Get the uid of the first key in the keyring.
+ *
+ * TODO: Set errno on failure.
+ * TODO: Check upstream calls to this function - they likely won't
+ *       handle the new error condition.
+ */
+int
+keyring_get_first_ring(pgp_keyring_t *ring, char *id, size_t len, int last)
+{
+    uint8_t	*src;
+
+    /* The NULL test on the ring may not be necessary for non-debug
+     * builds - it would be much better that a NULL ring never
+     * arrived here in the first place.
+     *
+     * The ring length check is a temporary fix for a case where
+     * an empty ring arrives and causes an access violation in
+     * some circumstances.
+     */
+
+    errno = 0;
+
+    if (ring == NULL || ring->keyc < 1) {
+        errno = EINVAL;
+        return 0;
+    }
+
+    memset(id, 0x0, len);
+
+    src = (uint8_t *) &ring->keys[(last) ? ring->keyc - 1 : 0].sigid;
+    keyring_format_key(id, src, len);
+
+    return 1;
+}
