+++ src/lib/keyring_pgp.c
@@ -92,7 +92,161 @@ __RCSID("$NetBSD: keyring.c,v 1.50 2011/06/25 00:37:44 agc Exp $");
 #include "rnpdigest.h"
 #include <json.h>
 
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <sys/param.h>
 
+#include <stdio.h>
+#include <string.h>
+#include <stdlib.h>
+
+/* read any gpg config file */
+static int
+conffile(rnp_t *rnp, char *homedir, char *userid, size_t length)
+{
+    regmatch_t	 matchv[10];
+    regex_t		 keyre;
+    char		 buf[BUFSIZ];
+    FILE		*fp;
+
+    __PGP_USED(rnp);
+    (void) snprintf(buf, sizeof(buf), "%s/gpg.conf", homedir);
+    if ((fp = fopen(buf, "r")) == NULL) {
+        return 0;
+    }
+    (void) memset(&keyre, 0x0, sizeof(keyre));
+    (void) regcomp(&keyre, "^[ \t]*default-key[ \t]+([0-9a-zA-F]+)",
+                   REG_EXTENDED);
+    while (fgets(buf, (int)sizeof(buf), fp) != NULL) {
+        if (regexec(&keyre, buf, 10, matchv, 0) == 0) {
+            (void) memcpy(userid, &buf[(int)matchv[1].rm_so],
+                          MIN((unsigned)(matchv[1].rm_eo -
+                                         matchv[1].rm_so), length));
+            if (rnp->passfp == NULL) {
+                (void) fprintf(stderr,
+                               "rnp: default key set to \"%.*s\"\n",
+                               (int)(matchv[1].rm_eo - matchv[1].rm_so),
+                               &buf[(int)matchv[1].rm_so]);
+            }
+        }
+    }
+    (void) fclose(fp);
+    regfree(&keyre);
+    return 1;
+}
+
+/* read a keyring and return it */
+static void *
+readkeyring(rnp_t *rnp, const char *name, const char *homedir)
+{
+    pgp_keyring_t	*keyring;
+    const unsigned	 noarmor = 0;
+    char		 f[MAXPATHLEN];
+    char		*filename;
+
+    if ((filename = rnp_getvar(rnp, name)) == NULL) {
+        (void) snprintf(f, sizeof(f), "%s/%s.gpg", homedir, name);
+        filename = f;
+    }
+    if ((keyring = calloc(1, sizeof(*keyring))) == NULL) {
+        (void) fprintf(stderr, "readkeyring: bad alloc\n");
+        return NULL;
+    }
+    if (!pgp_keyring_fileread(keyring, noarmor, filename)) {
+        free(keyring);
+        (void) fprintf(stderr, "cannot read %s %s\n", name, filename);
+        return NULL;
+    }
+    rnp_setvar(rnp, name, filename);
+    return keyring;
+}
+
+int
+pgp_keyring_load_keys(rnp_t *rnp, char *homedir)
+{
+	char     *userid;
+	char      id[MAX_ID_LENGTH];
+	pgp_io_t *io = rnp->io;
+
+	/* TODO: Some of this might be split up into sub-functions. */
+	/* TODO: Figure out what unhandled error is causing an
+	 *       empty keyring to end up in keyring_get_first_ring
+	 *       and ensure that it's not present in the
+	 *       ssh implementation too.
+	 */
+
+	rnp->pubring = readkeyring(rnp, "pubring", homedir);
+
+	if (rnp->pubring == NULL) {
+		fprintf(io->errs, "cannot read pub keyring\n");
+		return 0;
+	}
+
+	if (((pgp_keyring_t *) rnp->pubring)->keyc < 1) {
+		fprintf(io->errs, "pub keyring is empty\n");
+		return 0;
+	}
+
+	/* If a userid has been given, we'll use it. */
+	if ((userid = rnp_getvar(rnp, "userid")) == NULL) {
+		/* also search in config file for default id */
+		memset(id, 0, sizeof(id));
+		conffile(rnp, homedir, id, sizeof(id));
+		if (id[0] != 0x0) {
+			rnp_setvar(rnp, "userid", userid = id);
+		}
+	}
+
+	/* Only read secret keys if we need to. */
+	if (rnp_getvar(rnp, "need seckey")) {
+
+		rnp->secring = readkeyring(rnp, "secring", homedir);
+
+		if (rnp->secring == NULL) {
+			fprintf(io->errs, "cannot read sec keyring\n");
+			return 0;
+		}
+
+		if (((pgp_keyring_t *) rnp->secring)->keyc < 1) {
+			fprintf(io->errs, "sec keyring is empty\n");
+			return 0;
+		}
+
+		/* Now, if we don't have a valid user, use the first
+		 * in secring.
+		 */
+		if (! userid && rnp_getvar(rnp,
+								   "need userid") != NULL) {
+
+			if (!keyring_get_first_ring(rnp->secring, id,
+										sizeof(id), 0)) {
+				/* TODO: This is _temporary_. A more
+				 *       suitable replacement will be
+				 *       required.
+				 */
+				fprintf(io->errs, "failed to read id\n");
+				return 0;
+			}
+
+			rnp_setvar(rnp, "userid", userid = id);
+		}
+
+	} else if (rnp_getvar(rnp, "need userid") != NULL) {
+		/* encrypting - get first in pubring */
+		if (! userid && keyring_get_first_ring(
+				rnp->pubring, id, sizeof(id), 0)) {
+			rnp_setvar(rnp, "userid", userid = id);
+		}
+	}
+
+	if (! userid && rnp_getvar(rnp, "need userid")) {
+		/* if we don't have a user id, and we need one, fail */
+		fprintf(io->errs, "cannot find user id\n");
+		return 0;
+	}
+
+	return 1;
+}
 
 /**
    \ingroup HighLevel_Keyring
